backend/requirements.txt
beautifulsoup4==4.12.3
celery_progress==0.3
crewai==0.28.8
Django==5.0.3
djangorestframework==3.15.0
EbookLib==0.18
fitz==0.0.1.dev2
langchain==0.1.17
langchain_community==0.0.36
langchain_openai==0.1.5
ninja==1.11.1.1
pandas==2.2.2
pydantic==2.7.1
six==1.16.0
stripe==8.7.0
